requirements.txt
google-generativeai
google-genai
numpy~=2.2.5
pandas~=2.2.3
chromadb
langchain~=0.3.25
langchain-chroma
langchain-community
pydantic~=2.11.4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask~=3.1.1
fastapi~=0.115.9
uvicorn~=0.34.2
starlette~=0.45.3
python-multipart
unstructured